Break up buffer mixer and runner into modules
